requirements.txt
streamlit
tensorflow>=2.13,<2.16
opencv-python-headless
numpy
pandas
pillow
matplotlib
scikit-learn
